Resolves https://github.com/open-telemetry/opentelemetry-go-contrib/issues/4449.