@pellared 2023-11-24
@Aneurysm9 Would there be a problem if the IDs would conflict? May worry is not about the security but the possible insufficient entropy of the current implementation.

However, I agree that the current implementation can be acceptable at this point of time given the precedence of Trace and Span ID generation logic.

@takp Can you make some benchmarks to compare the implementations?