@takp 2023-11-24
Sure! I'll prepare benchmark.