@takp 2023-11-24
```go
package xray

import (
	"context"
	"testing"

	"go.opentelemetry.io/otel/trace"
)

// Benchmark for NewSpanID
func BenchmarkIDGenerator_NewSpanID(b *testing.B) {
	gen := NewIDGenerator()
	ctx := context.Background()
	traceID := trace.TraceID{}

	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		_ = gen.NewSpanID(ctx, traceID)
	}
}

// Benchmark for NewIDs
func BenchmarkIDGenerator_NewIDs(b *testing.B) {
	gen := NewIDGenerator()
	ctx := context.Background()

	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		_, _ = gen.NewIDs(ctx)
	}
}
````