@takp 2023-11-24
@Aneurysm9 @pellared I executed the benchmark and the original implementation is much faster with my local mac.

- Original implementation is approx. 45 times faster than the PR for `NewSpanID`.
- Original implementation is approx. 15 times faster than the PR for `NewIDs`.

I think this is not a trivial difference. If this module doesn't require cryptographical security, I think we don't have to change current implementation 👍 What do you think?

---

#### Original implementation:
```sh
$ go test -bench=.
goos: darwin
goarch: amd64
pkg: go.opentelemetry.io/contrib/propagators/aws/xray
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkIDGenerator_NewSpanID-12           	45713550	        26.06 ns/op
BenchmarkIDGenerator_NewIDs-12              	 6522184	       182.1 ns/op
BenchmarkStartAndEndSampledSpan-12          	 1804806	       666.0 ns/op
BenchmarkStartAndEndNestedSampledSpan-12    	 2208139	       542.4 ns/op
BenchmarkPropagatorExtract-12               	31596346	        37.83 ns/op
BenchmarkPropagatorInject-12                	51045776	        23.42 ns/op
PASS
ok  	go.opentelemetry.io/contrib/propagators/aws/xray	9.945s
```

#### This PR:
```sh
$ go test -bench=.
goos: darwin
goarch: amd64
pkg: go.opentelemetry.io/contrib/propagators/aws/xray
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkIDGenerator_NewSpanID-12           	 1027538	      1181 ns/op
BenchmarkIDGenerator_NewIDs-12              	  444454	      2690 ns/op
BenchmarkStartAndEndSampledSpan-12          	  363070	      3224 ns/op
BenchmarkStartAndEndNestedSampledSpan-12    	  666268	      1734 ns/op
BenchmarkPropagatorExtract-12               	30904088	        38.24 ns/op
BenchmarkPropagatorInject-12                	51253266	        23.43 ns/op
PASS
ok  	go.opentelemetry.io/contrib/propagators/aws/xray	8.525s
```


#### Benchmark

```go
package xray

import (
	"context"
	"testing"

	"go.opentelemetry.io/otel/trace"
)

// Benchmark for NewSpanID
func BenchmarkIDGenerator_NewSpanID(b *testing.B) {
	gen := NewIDGenerator()
	ctx := context.Background()
	traceID := trace.TraceID{}

	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		_ = gen.NewSpanID(ctx, traceID)
	}
}

// Benchmark for NewIDs
func BenchmarkIDGenerator_NewIDs(b *testing.B) {
	gen := NewIDGenerator()
	ctx := context.Background()

	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		_, _ = gen.NewIDs(ctx)
	}
}
````