@pellared 2023-11-24
@takp I agree. If the collisions would ever be a concern then the API may be extended to accept a custom ID generator. 

Closing. I will make a PR to get rid of the TODO comments next week. Thank you.