@Aneurysm9 2023-11-27
> the API may be extended to accept a custom ID generator.

This already is a custom ID generator that was needed only for compatibility with AWS X-Ray's former requirement that trace IDs start with a recent timestamp.  That requirement [has been eliminated](https://aws.amazon.com/about-aws/whats-new/2023/10/aws-x-ray-w3c-format-trace-ids-distributed-tracing/), but we need to keep this implementation here as it's in a stable module and it may still be used by existing systems for some time to come.  Not sure it's worth a major version bump to remove it but we can probably mark it as deprecated at this point.