@pellared 2023-11-28
I created https://github.com/open-telemetry/opentelemetry-go-contrib/pull/4635